(Chore) Edited  README.md and added.gitignore to root
The README.md file was edited to feature the travis badge, and also the
group app heading was made smaller to be on the same level with other heading
in the same category.

Other part of the README.md was also edited to make the title case, and .gitignore
added to remove vendore directory from pulling/pushing to github.